cargar foto de perfil

+++ controllers/post.loadPerfilePicture.controller.php
@@ -4,15 +4,14 @@ date_default_timezone_set('America/Bogota');
 header('Access-Control-Allow-Origin: *');
 
 require_once "models/connection.php";
-require_once "models/post.uploadDocumentsCustomer.model.php";
-require_once "models/put.updatePending.model.php";
+require_once "models/put.toUpdate.model.php";
 
 //include_once '../cors.php';
 
 class PostController
 {
     /*=============================================
-    Peticion post para crear cliente
+    Peticion post para crear foto de perfil del cliente
     =============================================*/
     static public function postRegister($table, $suffix, $id, $file)
     {
@@ -28,7 +27,7 @@ class PostController
 
         //$id_usuario = '99';
 
-        $target_path_nuevo = "documentos/" . $id_usuario . "/";
+        $target_path_nuevo = "src/perfile_pictures/clients/" . $id_usuario . "/";
         error_log("Nuevo Path: " . $target_path_nuevo);
 
         if (!file_exists("./" . $target_path_nuevo)) {
@@ -51,9 +50,56 @@ class PostController
             $return->fncResponse($response);
         } else {
 
-            $response = PostModel::postData("documentos", $suffix, $id_usuario, $target_path_nuevo);
+            $responsePendiente = PutModel::putDataPerfilePicture($table, $target_path_nuevo, "id_". $suffix, $id_usuario, $suffix);
 
-            $responsePendiente = PutPendingModel::putData($table, "id_" . $suffix, $id_usuario, $suffix);
+            move_uploaded_file($file['tmp_name'], "./" . $target_path_nuevo);
+
+            $return = new PostController();
+            $return->fncResponse($responsePendiente);
+        }
+    }
+
+    /*=============================================
+    Peticion post para crear foto de perfil del agente
+    =============================================*/
+    static public function postRegisterAgent($table, $suffix, $id, $file)
+    {
+
+        $target_path = "uploads/";
+        $target_path = $target_path . basename($file['name']);
+
+        error_log("Path: " . $target_path);
+
+
+        $nombreArchivo = $file['name'];
+        $id_usuario = $id;
+
+        //$id_usuario = '99';
+
+        $target_path_nuevo = "src/perfile_pictures/agents/" . $id_usuario . "/";
+        error_log("Nuevo Path: " . $target_path_nuevo);
+
+        if (!file_exists("./" . $target_path_nuevo)) {
+            if (mkdir("./" . $target_path_nuevo, 0777, true)) {
+                error_log("Exito! Carpeta creada:" . $target_path_nuevo);
+            } else {
+                error_log(" :( No pudo crear:" . $target_path_nuevo);
+            }
+        } else {
+            error_log("Carpeta existente:" . $target_path_nuevo);
+        }
+
+        $target_path_nuevo = $target_path_nuevo . $nombreArchivo;
+
+        if (file_exists("./" . $target_path_nuevo)) {
+            $response = array(
+                "code" => 13
+            );
+            $return = new PostController();
+            $return->fncResponse($response);
+        } else {
+
+            $responsePendiente = PutModel::putDataPerfilePicture($table, $target_path_nuevo, "id_". $suffix, $id_usuario, $suffix);
 
             move_uploaded_file($file['tmp_name'], "./" . $target_path_nuevo);
